cleaned up remove
