+++ Railgun/Commands/Music/MusicRoleLock.cs
@@ -11,7 +11,7 @@ namespace Railgun.Commands.Music
     public partial class Music
 	{
 		[Alias("rolelock"), UserPerms(GuildPermission.ManageGuild)]
-		public class MusicRoleLock : SystemBase
+		public partial class MusicRoleLock : SystemBase
 		{
 			[Command("add")]
 			public Task AddRoleAsync(IRole role)
@@ -31,33 +31,6 @@ namespace Railgun.Commands.Music
 				foreach (var role in Context.Guild.Roles)
 					if (role.Name.Contains(roleName))
 						return AddRoleAsync(role);
-				return ReplyAsync("Unable to find a role with the name you specified.");
-			}
-
-			[Command("remove")]
-			public Task RemoveRoleAsync(IRole role)
-			{
-				var profile = Context.Database.ServerProfiles.GetOrCreateData(Context.Guild.Id);
-            	var data = profile.Music;
-				var removed = data.RemoveAllowedRole(role.Id);
-
-				if (!removed)
-					return ReplyAsync("The role specified was never role-locked.");
-
-				var output = new StringBuilder()
-					.AppendFormat("The role {0}, has now been removed from role-locking.", role.Name).AppendLine();
-
-				if (data.AllowedRoles.Count < 1) output.AppendLine("All music commands are no longer role-locked.");
-
-				return ReplyAsync(output.ToString());
-			}
-			
-			[Command("remove")]
-			public Task RemoveRoleAsync([Remainder] string roleName)
-			{
-				foreach (var role in Context.Guild.Roles)
-					if (role.Name.Contains(roleName))
-						return RemoveRoleAsync(role);
 				return ReplyAsync("Unable to find a role with the name you specified.");
 			}
 		}
